+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Enhancements:**
 - Support for Stimulus 1.0
+- `import`s are now sorted alphabetically every time you run the component generator
 
 ## v1.1.1 (2018-01-20)